refactor(events): extract check-in flow into check_in_service
ticket_service.py sat at exactly the 1000-line ceiling enforced by
`make check`, so any addition to it fails CI. Move the door check-in
cluster verbatim into events/service/check_in_service.py:
_format_in_event_tz, _check_in_closed_message, resolve_check_in_ticket_id,
_pending_check_in_allowed and check_in_ticket.

Repoints member_scan_service, the event-admin tickets controller, their
tests and the doc comments that name them. No behaviour change;
ticket_service drops to 850 lines.
