+++ src/events/utils/__init__.py
@@ -421,7 +421,7 @@ def create_series_pass_pdf(held_pass: "HeldSeriesPass") -> bytes:
     organization = event_series.organization
 
     # held_pass.qr_payload is the single source of truth for the check-in contract
-    # (see ticket_service.resolve_check_in_ticket_id).
+    # (see check_in_service.resolve_check_in_ticket_id).
     qr_code_base64 = _qr_code_base64(held_pass.qr_payload)
 
     links = list(series_pass.tier_links.select_related("event").order_by("event__start"))
@@ -481,7 +481,7 @@ def create_membership_pdf(member: "OrganizationMember") -> bytes:
     organization = member.organization
 
     # member.qr_payload is the single source of truth for the scan contract
-    # (see ticket_service.resolve_check_in_ticket_id / member_scan_service.scan_member_code).
+    # (see check_in_service.resolve_check_in_ticket_id / member_scan_service.scan_member_code).
     qr_code_base64 = _qr_code_base64(member.qr_payload)
 
     logo_file = organization.logo_thumbnail or organization.logo
